stdio: Begin printf family of functions
Adds the <stdio.h> header and starts implementing the printf family.
Most of the printf functions are defined in terms of one another,
eventually boiling down to a callback-based printf implementation.
snprintf() is implemented using this, while the FILE-based printf
functions are not yet added.

The initial implementation of the callback-based printf supports
formatting the % sign, characters, and strings, with padding and
precision flags.